Add full server manager panel and fix dashboard challenge
