+++ .env
@@ -2,3 +2,4 @@ NEXT_PUBLIC_API_BASEURL=http://127.0.0.1:5000/api/v2/
 NEXT_PUBLIC_SOCKET=http://127.0.0.1:5000
 NEXT_PUBLIC_HOSTNAME=http://localhost:3000
 NEXT_PUBLIC_SHOW_SPONSORS=false
+NEXT_PUBLIC_SERVICE_MANAGE_PANEL=fullserver
