+++ src/components/module/dashboard/service/interface.ts
@@ -1,0 +1,6 @@
+import { Challenge } from "@/types/challenge";
+
+export interface TeamChallServiceProps {
+    chall: Challenge | undefined;
+    isUnlocked: boolean;
+}
